requirements.txt
# PyQt6 Framework
PyQt6==6.10.0
PyQt6-Qt6==6.10.0
PyQt6-sip==13.6.0

# Security
bcrypt==4.1.2

# Utilities
python-dateutil==2.8.2
pytz==2024.1
python-dotenv==1.0.0

# UI & Icons
QtAwesome==1.3.1

# Data Export
openpyxl==3.1.2
matplotlib==3.8.2

# Task Scheduling
APScheduler==3.10.4

# Development (optional)
# pytest==8.0.0
# pytest-qt==4.3.1
# pylint==3.0.3
